refactor(twitter): delete browser CDP client and session modules
